third_party/blink/web_tests/platform/mac/external/wpt/service-workers/service-worker/navigation-timing-extended.https-expected.txt

This is a testharness.js-based test.
FAIL Service worker controlled navigation timing assert_greater_than: workerStart marking should not wait for worker activation to finish expected a number greater than 1634043789619.9001 but got 1634043789618.7002
Harness: the test ran to completion.
